[CRT-132] Rename projects to iframe-rpc
