+++ apps/scratch/package.json
@@ -37,7 +37,6 @@
     "@turbowarp/types": "^0.0.13",
     "@typescript-eslint/eslint-plugin": "^8.29.0",
     "@typescript-eslint/parser": "^8.29.0",
-    "app-iframe-message-react": "portal:../../libraries/app-iframe-message-react",
     "arraybuffer-loader": "^1.0.8",
     "autoprefixer": "^10.4.21",
     "babel-plugin-istanbul": "^7.0.0",
@@ -56,6 +55,7 @@
     "eslint-plugin-prettier": "^5.2.6",
     "get-float-time-domain-data": "^0.1.0",
     "get-user-media-promise": "1.1.4",
+    "iframe-rpc-react": "portal:../../libraries/iframe-rpc-react",
     "immutable": "^5.1.1",
     "js-base64": "^3.7.7",
     "jszip": "^3.10.1",
