+++ frontend/package.json
@@ -50,11 +50,11 @@
     "@fortawesome/react-fontawesome": "^0.2.2",
     "@hookform/resolvers": "^4.1.3",
     "@sentry/nextjs": "^9.13.0",
-    "app-iframe-message-react": "portal:../libraries/app-iframe-message-react",
     "bootstrap": "^5.3.3",
     "chart.js": "^4.4.8",
     "chartjs-plugin-annotation": "^3.1.0",
     "edit-distance": "^1.0.5",
+    "iframe-rpc-react": "portal:../libraries/iframe-rpc-react",
     "jqgram": "^0.2.5",
     "next": "15.3.1",
     "openid-client": "^6.3.4",
