+++ frontend/src/components/dashboard/StudentProgress.tsx
@@ -1,7 +1,7 @@
 import { Accordion } from "react-bootstrap";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { FormattedMessage, useIntl } from "react-intl";
-import { Language } from "app-iframe-message-react/src";
+import { Language } from "iframe-rpc-react/src";
 import { ExistingSessionExtended } from "@/api/collimator/models/sessions/existing-session-extended";
 import { ExistingClassExtended } from "@/api/collimator/models/classes/existing-class-extended";
 import { SessionTask } from "@/api/collimator/models/sessions/session-task";
@@ -12,9 +12,9 @@ import { useAllSessionSolutions } from "@/api/collimator/hooks/solutions/useAllS
 import { useFileHash } from "@/hooks/useFileHash";
 import { ExistingStudentSolution } from "@/api/collimator/models/solutions/existing-student-solutions";
 import { useSolutionFile } from "@/api/collimator/hooks/solutions/useSolution";
-import EmbeddedApp, { EmbeddedAppRef } from "../EmbeddedApp";
-import MultiSwrContent from "../MultiSwrContent";
 import SwrContent from "../SwrContent";
+import MultiSwrContent from "../MultiSwrContent";
+import EmbeddedApp, { EmbeddedAppRef } from "../EmbeddedApp";
 
 type Progress = ExistingStudentSolution;
 type ProgressByTask = { [taskId: number]: Progress };
