+++ libraries/iframe-rpc-react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "app-iframe-message-react",
+  "name": "iframe-rpc-react",
   "packageManager": "yarn@4.9.1",
   "files": [
     "src/**/*"
@@ -8,7 +8,7 @@
     "lint": "eslint ."
   },
   "dependencies": {
-    "app-iframe-message": "portal:../app-iframe-message"
+    "iframe-rpc": "portal:../iframe-rpc"
   },
   "peerDependencies": {
     "react": "^19.1.0"
