+++ libraries/iframe-rpc-react/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./useIframeParent";
 export * from "./useIframeChild";
 
-export * from "app-iframe-message/src";
+export * from "iframe-rpc/src";
